@@ -1,53 +1,35 @@
 {
-  "name": "dev.mentor-react-workshop",
-  "author": "Vitalijus Zudenkovas",
-  "version": "1.0.0",
+  "name": "fronted",
   "private": true,
-  "dependencies": {
-    "@mui/lab": "5.0.0-alpha.82",
-    "@mui/material": "5.8.0",
-    "@mui/styled-engine-sc": "5.8.0",
-    "@testing-library/jest-dom": "5.16.4",
-    "@testing-library/react": "13.2.0",
-    "@testing-library/user-event": "14.2.0",
-    "axios": "0.27.2",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
-    "react-query": "3.39.0",
-    "react-router-dom": "6.3.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "5.3.5"
-  },
-  "resolutions": {
-    "react-error-overlay": "6.0.9"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "tsc": "./node_modules/.bin/tsc",
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@emotion/styled": "^11.10.4",
+    "@hookform/resolvers": "^2.9.9",
+    "@mui/lab": "^5.0.0-alpha.103",
+    "@mui/material": "^5.10.9",
+    "@mui/styled-engine": "^5.10.8",
+    "@mui/styled-engine-sc": "^5.10.6",
+    "axios": "^1.1.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.37.0",
+    "react-query": "^3.39.2",
+    "react-router-dom": "^6.4.2",
+    "styled-components": "^5.3.6",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@craco/craco": "6.4.3",
-    "@types/node": "17.0.35",
-    "@types/react": "18.0.9",
-    "@types/react-dom": "18.0.4",
-    "@types/styled-components": "5.1.25",
-    "craco-alias": "3.0.1",
-    "react-error-overlay": "6.0.9",
-    "typescript": "4.6.4"
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@types/styled-components": "^5.1.26",
+    "@vitejs/plugin-react": "^2.1.0",
+    "typescript": "^4.6.4",
+    "vite": "^3.1.0"
   }
 }
